+correct type of X in Point.java

+++ back-end/src/Point.java
@@ -13,7 +13,7 @@ public class Point {
     @Column(name = "id", nullable = false)
     private Integer id;
     @Column(name = "x", nullable = false)
-    private Integer x;
+    private Double x;
     @Column(name = "y", nullable = false)
     private Double y;
     @Column(name = "r", nullable = false)
@@ -21,13 +21,13 @@ public class Point {
     @Column(name = "result", nullable = false)
     private Boolean result;
 
-    public Point (Integer x, Double y, Integer r) {
+    public Point (Double x, Double y, Integer r) {
         this.x = x;
         this.y = y;
         this.r = r;
     }
 
-    public Point (Integer x, Double y, Integer r, Boolean result) {
+    public Point (Double x, Double y, Integer r, Boolean result) {
         this.x = x;
         this.y = y;
         this.r = r;
@@ -42,11 +42,11 @@ public class Point {
         this.result = result;
     }
 
-    public Integer getX ( ) {
+    public Double getX ( ) {
         return x;
     }
 
-    public void setX (Integer x) {
+    public void setX (Double x) {
         this.x = x;
     }
